+++ src/pages/Preferences/Location.vue
@@ -3,6 +3,16 @@
     <div class="container">
       <h3>Aquí van la localización del usuario</h3>
       <br>
+      <l-map
+        style="height: 80%; width: 100%"
+        :zoom="zoom"
+        :center="center"
+        @updateate:zoom="zoomUpdated"
+        @update:center="centerUpdated"
+        @update:bounds="boundsUpdated"
+      >
+        <l-tile-layer :url="url"></l-tile-layer>
+      </l-map>
       <div class="preferencesButton">
         <v-ons-button
           class="preferencesButton__button"
@@ -19,14 +29,31 @@
 
 <script>
 import { mapMutations } from 'vuex'
+import { LMap, LTileLayer, LMarker } from 'vue2-leaflet'
 import Preferences from './Preferences'
+
 export default {
   name: 'location',
   data() {
-    return {}
+    return {
+      url: 'http://{s}.tile.osm.org/{z}/{x}/{y}.png',
+      zoom: 3,
+      center: [47.41322, -1.219482],
+      bounds: null
+    }
   },
   methods: {
     ...mapMutations('navigator', ['PUSH']),
+
+    zoomUpdated(zoom) {
+      this.zoom = zoom
+    },
+    centerUpdated(center) {
+      this.center = center
+    },
+    boundsUpdated(bounds) {
+      this.bounds = bounds
+    },
 
     toPreferences() {
       this.PUSH(Preferences)