+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2019-05-14
+
+### Changed
+
+- Refactored and modulizated the store
+
+### Added
+
+- User init preferences pages
+- Leaflet Maps
+
 ## [0.1.4] - 2019-05-09
 
 ### Changed